Migrate BaseController
