+++ common/src/main/java/com/microsoft/identity/common/internal/commands/DeviceCodeFlowCommand.java
@@ -25,7 +25,7 @@ package com.microsoft.identity.common.internal.commands;
 import androidx.annotation.NonNull;
 
 import com.microsoft.identity.common.java.WarningType;
-import com.microsoft.identity.common.internal.commands.parameters.DeviceCodeFlowCommandParameters;
+import com.microsoft.identity.common.java.commands.parameters.DeviceCodeFlowCommandParameters;
 import com.microsoft.identity.common.internal.controllers.BaseController;
 import com.microsoft.identity.common.internal.result.AcquireTokenResult;
 import com.microsoft.identity.common.java.providers.microsoft.microsoftsts.MicrosoftStsAuthorizationResponse;
