+++ common/src/main/java/com/microsoft/identity/common/internal/commands/RemoveCurrentAccountCommand.java
@@ -25,7 +25,7 @@ package com.microsoft.identity.common.internal.commands;
 import androidx.annotation.NonNull;
 
 import com.microsoft.identity.common.java.WarningType;
-import com.microsoft.identity.common.internal.commands.parameters.RemoveAccountCommandParameters;
+import com.microsoft.identity.common.java.commands.parameters.RemoveAccountCommandParameters;
 import com.microsoft.identity.common.internal.controllers.BaseController;
 
 import java.util.List;
