+++ common/src/main/java/com/microsoft/identity/common/internal/result/AdalBrokerResultAdapter.java
@@ -39,6 +39,7 @@ import com.microsoft.identity.common.java.exception.ErrorStrings;
 import com.microsoft.identity.common.java.exception.IntuneAppProtectionPolicyRequiredException;
 import com.microsoft.identity.common.java.exception.ServiceException;
 import com.microsoft.identity.common.java.exception.UserCancelException;
+import com.microsoft.identity.common.java.result.ILocalAuthenticationResult;
 import com.microsoft.identity.common.java.util.SchemaUtil;
 import com.microsoft.identity.common.java.dto.IAccountRecord;
 import com.microsoft.identity.common.logging.Logger;
@@ -151,7 +152,8 @@ public class AdalBrokerResultAdapter implements IBrokerResultAdapter {
     }
 
     @Override
-    public @NonNull ILocalAuthenticationResult authenticationResultFromBundle(Bundle resultBundle) {
+    public @NonNull
+    ILocalAuthenticationResult authenticationResultFromBundle(Bundle resultBundle) {
         throw new UnsupportedOperationException();
     }
 
@@ -296,8 +298,8 @@ public class AdalBrokerResultAdapter implements IBrokerResultAdapter {
         }
 
         //INTERACTION_REQUIRED is marked as deprecated
-        if (AuthenticationConstants.OAuth2ErrorCode.INVALID_GRANT.equalsIgnoreCase(serviceException.getErrorCode())
-                || AuthenticationConstants.OAuth2ErrorCode.INTERACTION_REQUIRED.equalsIgnoreCase(serviceException.getErrorCode())
+        if (com.microsoft.identity.common.java.AuthenticationConstants.OAuth2ErrorCode.INVALID_GRANT.equalsIgnoreCase(serviceException.getErrorCode())
+                || com.microsoft.identity.common.java.AuthenticationConstants.OAuth2ErrorCode.INTERACTION_REQUIRED.equalsIgnoreCase(serviceException.getErrorCode())
         ) {
 
             resultBundle.putString(
