+++ common/src/test/java/com/microsoft/identity/common/internal/commands/parameters/GenerateShrCommandParametersTest.java
@@ -28,6 +28,7 @@ import com.microsoft.identity.common.AndroidPlatformComponents;
 import com.microsoft.identity.common.java.authscheme.PopAuthenticationSchemeInternal;
 import com.microsoft.identity.common.internal.commands.CommandCallback;
 import com.microsoft.identity.common.java.commands.parameters.CommandParameters;
+import com.microsoft.identity.common.java.commands.parameters.GenerateShrCommandParameters;
 import com.microsoft.identity.common.java.util.ClockSkewManager;
 import com.microsoft.identity.common.java.util.UrlUtil;
 import com.microsoft.identity.common.java.util.ported.InMemoryStorage;
