+++ common4j/src/main/com/microsoft/identity/common/internal/result/AcquireTokenResult.java
@@ -25,6 +25,7 @@ package com.microsoft.identity.common.internal.result;
 import com.microsoft.identity.common.java.WarningType;
 import com.microsoft.identity.common.java.providers.oauth2.TokenResult;
 import com.microsoft.identity.common.java.providers.oauth2.AuthorizationResult;
+import com.microsoft.identity.common.java.result.ILocalAuthenticationResult;
 
 public class AcquireTokenResult {
