+++ common4j/src/main/com/microsoft/identity/common/java/adal/ADALUserInfo.java
@@ -20,11 +20,11 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.adal.internal;
+package com.microsoft.identity.common.java.adal;
 
-import com.microsoft.identity.common.adal.internal.util.DateExtensions;
 import com.microsoft.identity.common.java.providers.microsoft.azureactivedirectory.AzureActiveDirectoryAccount;
-import com.microsoft.identity.common.internal.result.ILocalAuthenticationResult;
+import com.microsoft.identity.common.java.result.ILocalAuthenticationResult;
+import com.microsoft.identity.common.java.util.DateExtensions;
 import com.microsoft.identity.common.java.util.SchemaUtil;
 
 import java.net.URL;
