+++ common4j/src/main/com/microsoft/identity/common/java/commands/parameters/DeviceCodeFlowCommandParameters.java
@@ -20,7 +20,7 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.common.internal.commands.parameters;
+package com.microsoft.identity.common.java.commands.parameters;
 
 import com.microsoft.identity.common.java.commands.parameters.TokenCommandParameters;
