+++ common4j/src/main/com/microsoft/identity/common/java/util/DateExtensions.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.adal.internal.util;
+package com.microsoft.identity.common.java.util;
 
 import java.util.Date;
