+++ common4j/src/main/com/microsoft/identity/common/java/util/StringUtil.java
@@ -35,10 +35,12 @@ import java.text.SimpleDateFormat;
 import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.StringTokenizer;
 import java.util.TimeZone;
 import java.util.UUID;
@@ -350,5 +352,36 @@ public class StringUtil {
         String outValue = null == input ? "" : input.toLowerCase(Locale.US).trim();
 
         return outValue;
+    }
+
+    /**
+     * This is a reimplementation of String.join for the android platform.  Possibly this should
+     * shift into PlatformUtils, which could rely on String.join dependent on the android API level.
+     * @param separator a separator for the joined strings.
+     * @param segments a set of segments to join.
+     * @return a new char sequence constructed by joining the segments with the separator.
+     */
+    public static <T extends CharSequence> String join(final CharSequence separator, final @NonNull Iterable<T> segments) {
+        final Iterator<T> itr = segments.iterator();
+        // If the iterable is empty, return empty string.
+        if (!itr.hasNext()) {
+            return "";
+        }
+        final T first = itr.next();
+        // If the iterable has but one value, return it directly.
+        if (!itr.hasNext()) {
+            if (first instanceof String) {
+                return (String) first;
+            }
+            return first.toString();
+        }
+        final StringBuilder sb = new StringBuilder();
+        // This iterator must have at least one value, since it isn't empty.
+        sb.append(first);
+        while (itr.hasNext()) {
+            sb.append(separator);
+            sb.append(itr.next());
+        }
+        return sb.toString();
     }
 }
